Add files via upload
This repository will serve as the central hub for our final project's development and collaboration. This repository contains the datasets and files utilized in our project, which centered around assessing trust and confidence through the analysis of provenance and lineage. Our specific focus was on news sources within the domain. Given the multitude of news outlets reporting on the same events, our objective was to establish a confidence score for each source. Leveraging provenance and lineage, we aimed to identify the most reliable news sources in order to enhance the credibility assessment of reported events.